RFC #4 -- Adjusted the directory layout to facilitate support of multiple compilers.  please NOTE: the location of the Visual Studio solution/project files, as well as the resulting object library files has changed.
Currently we support two compilers of the Microsoft variety: VS-2008 (version 9.0) and VS-2010 (version 10.0).  There is virtually no code that needs changing between the two compilers, but all the solution and project files do need to change in order to operate with the new IDE.  (Yes, it is the IDE that is different.  The underlying compiler (cl.exe) doesn't change much at all.).

In addition, I took this RFC coase to eliminate all W4 warnings from the existing code base.  So the product should now compile clean and the Console Test CPP application should run clean.  Next job is to get rid of all gcc compiler warning messages.

Finally, the Code Review accomplished before submission mentioned some inconsistencies concerning the use of slash vs backslash in the new project files, and also the in consistent use of the executable names of the Console test modules.  I'll deal with these in another submission, and justify the submission as being part of RFC 4.  Need to get the VS 2010 feature supported so that other OsGeo projects can get rolling on their efforts to port to VS 2010.

git-svn-id: https://svn.osgeo.org/metacrs/csmap/trunk/CsMapDev@1991 4e78687f-474d-0410-85f9-8d5e500ac6b2

+++ CsMapConsoleUtilities/Source/csUtilitySupport.cpp
@@ -124,11 +124,11 @@ bool CS_strrpl (char* string1,size_t strSize,const char* find,const char* rplWit
 			count3 = strlen (rplWith);
 			count4 = strlen (string1) - count1 - count2;
 
-			cPtr = CS_stncp (wrkStr,string1,count1 + 1);
-			cPtr = CS_stncp (cPtr,rplWith,count3 + 1);
-			cPtr = CS_stncp (cPtr,(string1 + count1 + count2),count4 + 1);
+			cPtr = CS_stncp (wrkStr,string1,(int)(count1 + 1));
+			cPtr = CS_stncp (cPtr,rplWith,(int)(count3 + 1));
+			cPtr = CS_stncp (cPtr,(string1 + count1 + count2),(int)(count4 + 1));
 			
-			CS_stncp (string1,wrkStr,strSize);
+			CS_stncp (string1,wrkStr,(int)strSize);
 			ok = true;
 		}
 	}
